Move shared preferences module to module package
